pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Backend for Приложение КИТЭКа"
authors = [
    {name = "xaoriGit",email = "xaorinit@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "django-ninja (>=1.4.3,<2.0.0)",
    "django (>=5.2.4,<6.0.0)",
    "django-admin-interface (>=0.30.1,<0.31.0)",
    "django-colorfield (>=0.14.0,<0.15.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "firebase-admin (>=6.9.0,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "django-ninja-jwt (>=5.3.7,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-stubs = "^5.2.1"

